requirements.txt
annotated-types==0.6.0
anyio==4.2.0
argcomplete==3.2.2
black==24.2.0
boto3==1.34.56
botocore==1.34.56
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
commitizen==3.16.0
decli==0.6.1
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.36
pyaes==1.6.1
pyasn1==0.5.1
pycryptodome==3.20.0
pydantic==2.6.1
pydantic_core==2.16.2
pytest==8.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
questionary==2.0.1
requests==2.31.0
rsa==4.9
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
termcolor==2.4.0
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.9.0
urllib3==1.26.18
wcwidth==0.2.13
zipp==3.17.0
